revove `pending` invoice
