+++ app/jobs/bill_non_invoiceable_fees_job.rb
@@ -6,7 +6,7 @@ class BillNonInvoiceableFeesJob < ApplicationJob
   retry_on Sequenced::SequenceError, ActiveJob::DeserializationError
 
   def perform(subscriptions, billing_at)
-    result = Invoices::RegroupFeesService.call(subscriptions:, billing_at:)
+    result = Invoices::AdvanceChargesService.call(subscriptions:, billing_at:)
     result.raise_if_error!
   end
 end
